More changes to __mangling for classes
Now includes `cdef` classes. Also undoes the changes to Visitor.py
Combines the tests (mostly) in one file
