+++ tests/run/methodmangling_cdef.pxd
@@ -1,0 +1,3 @@
+cdef class InPxd:
+    cdef public int __y
+    cdef int __private_cdef(self)
