--- tests/run/methodmangling_pure_T1382.pxd
@@ -1,2 +1,0 @@
-cdef class CDefClassInPxd:
-    pass
